Split sanitization from data API calls
