+++ lib/community/communityFields.js
@@ -13,6 +13,22 @@ import {
 
 import { EventType } from '../event/eventFields';
 
+function eventsList(communityProperty) {
+  return {
+    type: new GraphQLList(EventType),
+    description: 'The community\'s events',
+    resolve(community) {
+      if (community) {
+        return community[communityProperty].map(result => {
+          const id = pathOr('', ['data', 'value', 'document', 'id'], result);
+          return getDocumentById(id, 'events', sanitizeEventAndNews);
+        });
+      }
+      return [];
+    },
+  };
+}
+
 const communityType = new GraphQLObjectType({
   name: 'Community',
   fields: {
@@ -47,19 +63,7 @@ const communityType = new GraphQLObjectType({
       type: GraphQLString,
       description: 'The summary of the community mailing list',
     },
-    events: {
-      type: new GraphQLList(EventType),
-      description: 'The community\'s events',
-      resolve(community) {
-        if (community) {
-          return community.events.map((result) => {
-            const id = pathOr('', ['data', 'value', 'document', 'id'], result);
-            return getDocumentById(id, 'event', sanitizeEventAndNews);
-          });
-        }
-        return [];
-      },
-    },
+    events: eventsList('events'),
   },
 });
