+++ lib/fetch/sanitize.spec.js
@@ -1,67 +1,13 @@
-import PrismicMock from '../test/mocks/prismic';
 import * as prismicEventResponse from
-  '../test/fixtures/event-prismic-response.json';
+  '../../test/fixtures/event-prismic-response.json';
 
 import {
   sanitizeEventAndNews,
-  fetchEvent,
-  fetchAllEvents,
-  fetchNewsArticle,
-  fetchAllNews,
   sanitizeCommunity,
   sanitizeSpeaker,
   sanitizeTalk,
   reformatLinkList,
-} from './fetch';
-
-describe('Fetch', () => {
-  let prismicMock;
-
-  beforeEach(() => {
-    prismicMock = new PrismicMock('https://rb-website-stage.prismic.io');
-    prismicMock.mockApi();
-  });
-
-  describe('All', () => {
-    it('should fetch events', async () => {
-      const docType = 'event';
-      const scope = prismicMock.mockDocumentTypeQuery(docType);
-
-      await fetchAllEvents();
-
-      expect(scope.isDone()).to.equal(true);
-    });
-
-    it('should fetch news', async () => {
-      const docType = 'news';
-      const scope = prismicMock.mockDocumentTypeQuery(docType);
-
-      await fetchAllNews();
-
-      expect(scope.isDone()).to.equal(true);
-    });
-  });
-
-  describe('Individual', () => {
-    it('should fetch an individual event', async () => {
-      const id = 'falafels';
-      const scope = prismicMock.mockDocumentIdQuery(id);
-
-      await fetchEvent(undefined, { id }, { headers: {} });
-
-      expect(scope.isDone()).to.equal(true);
-    });
-
-    it('should fetch an individual news article', async () => {
-      const id = 'falafels';
-      const scope = prismicMock.mockDocumentIdQuery(id);
-
-      await fetchNewsArticle(undefined, { id }, { headers: {} });
-
-      expect(scope.isDone()).to.equal(true);
-    });
-  });
-});
+} from './sanitize';
 
 describe('Data sanitation', () => {
   it('should sanitate empty fields', () => {
